+++ mods/src/server/server.cpp
@@ -61,6 +61,7 @@ static ServerProperties &get_server_properties() {
 #define DEFAULT_MAX_PLAYERS "10"
 #define DEFAULT_WHITELIST "false"
 #define DEFAULT_DEATH_MESSAGES "true"
+#define DEFAULT_GENERATE_CAVES "true"
 
 // Get World Name
 static std::string get_world_name() {
@@ -493,6 +494,9 @@ static const char *get_features() {
         if (get_server_properties().get_bool("death-messages", DEFAULT_DEATH_MESSAGES)) {
             features += "Implement Death Messages|";
         }
+        if (get_server_properties().get_bool("generate-caves", DEFAULT_GENERATE_CAVES)) {
+            features += "Generate Caves|";
+        }
     }
     return features.c_str();
 }
@@ -542,6 +546,8 @@ static void server_init() {
         properties_file_output << "whitelist=" DEFAULT_WHITELIST "\n";
         properties_file_output << "# Enable Death Messages\n";
         properties_file_output << "death-messages=" DEFAULT_DEATH_MESSAGES "\n";
+        properties_file_output << "# Generate Caves\n";
+        properties_file_output << "generate-caves=" DEFAULT_GENERATE_CAVES "\n";
         properties_file_output.close();
         // Re-Open File
         properties_file = std::ifstream(file);