+++ symbols/include/symbols/minecraft.h
@@ -76,6 +76,8 @@ static unsigned char *Options_Option_ANAGLYPH = (unsigned char *) 0x136c08; // O
 
 static bool *Minecraft_useAmbientOcclusion = (bool *) 0x136b90;
 
+static bool *HeavyTile_instaFall = (bool *) 0x180cc0;
+
 // Structures
 
 struct AABB {
@@ -129,8 +131,20 @@ static uint32_t Tile_setSoundType_vtable_offset = 0xe8;
 typedef int32_t (*Tile_use_t)(unsigned char *tile, unsigned char *level, int32_t x, int32_t y, int32_t z, unsigned char *player);
 static uint32_t Tile_use_vtable_offset = 0x98;
 
+typedef int32_t (*Tile_getColor_t)(unsigned char *tile, unsigned char *level_source, int32_t x, int32_t y, int32_t z);
+static uint32_t Tile_getColor_vtable_offset = 0xb8;
+
 static uint32_t Tile_id_property_offset = 0x8; // int32_t
 static uint32_t Tile_category_property_offset = 0x3c; // int32_t
+
+// GrassTile
+
+static void *GrassTile_getColor_vtable_addr = (void *) 0x111660;
+
+// TallGrass
+
+static Tile_getColor_t TallGrass_getColor = (Tile_getColor_t) 0xc25fc;
+static void *TallGrass_getColor_vtable_addr = (void *) 0x1121f0;
 
 // TileRenderer
 
@@ -511,6 +525,11 @@ static Level_clip_t Level_clip = (Level_clip_t) 0xa3db0;
 
 static uint32_t Level_players_property_offset = 0x60; // std::vector<ServerPlayer *>
 
+// LevelSource
+
+typedef unsigned char *(*LevelSource_getBiome_t)(unsigned char *level_source, int32_t x, int32_t z);
+static uint32_t LevelSource_getBiome_vtable_offset = 0x24;
+
 // Material
 
 typedef bool (*Material_isSolid_t)(unsigned char *material);
@@ -858,6 +877,35 @@ static uint32_t HumanoidModel_is_sneaking_property_offset = 0x236; // bool
 // PlayerRenderer
 
 static void *PlayerRenderer_render_vtable_addr = (void *) 0x107f08;
+
+// WorkbenchScreen
+
+#define WORKBENCH_SCREEN_SIZE 0x180
+
+typedef unsigned char *(*WorkbenchScreen_t)(unsigned char *screen, int32_t param_1);
+static WorkbenchScreen_t WorkbenchScreen = (WorkbenchScreen_t) 0x301cc;
+
+// Biome
+
+static uint32_t Biome_color_property_offset = 0x2c; // int32_t
+static uint32_t Biome_leaf_color_property_offset = 0x34; // int32_t
+
+// LargeFeature
+
+typedef void (*LargeFeature_apply_t)(unsigned char *large_feature, unsigned char *chunk_source, unsigned char *level, int32_t chunk_x, int32_t chunk_y, unsigned char *chunk_data, int32_t unused);
+static uint32_t LargeFeature_apply_vtable_offset = 0x8;
+
+// RandomLevelSource
+
+typedef void (*RandomLevelSource_postProcess_t)(unsigned char *random_level_source, unsigned char *chunk_source, int32_t chunk_x, int32_t chunk_y);
+static RandomLevelSource_postProcess_t RandomLevelSource_postProcess = (RandomLevelSource_postProcess_t) 0xb2238;
+static void *RandomLevelSource_postProcess_vtable_addr = (void *) 0x1105ac;
+
+typedef void (*RandomLevelSource_buildSurface_t)(unsigned char *random_level_source, int32_t chunk_x, int32_t chunk_y, unsigned char *chunk_data, unsigned char **biomes);
+static RandomLevelSource_buildSurface_t RandomLevelSource_buildSurface = (RandomLevelSource_buildSurface_t) 0xb32ec;
+
+static uint32_t RandomLevelSource_cave_feature_property_offset = 0x8; // LargeCaveFeature
+static uint32_t RandomLevelSource_level_property_offset = 0x72c8; // Level *
 
 // Method That Require C++ Types
 #ifdef __cplusplus